Ch 3 Task 3 completed
Write a program that calculates the sum of numbers entered by the user. The program asks for a number, reads the number entered by the user, adds it to the sum, and asks for the number again. The process continues until the user enters an  empty value. Use exception handling