Ch 3 Task 4 completed
Write a program where user enters an integer in the range from 1 to 7, and the program determines the day of the week from this number. If the number is out of range, display a message that the entered value is invalid. Use a switch statement. Suggest a mechanism for handling an error associated with entering a non-numeric value.