Ch 4 Task 2 completed
Write a program that creates an array of 11 integers. The array is filled with powers of two. When filling the array, keep in mind that the initial element is equal to 1, and each subsequent element is 2 times greater than the previous one. Display the array in the console window in forward and reverse order. The size of the array is given by a variable.