Ch 4 Task 5 completed
Write a program that creates an array and fills it with random numbers. The array is displayed in the console. In this array, you need to determine the element with the minimum value. In particular, the program should output the value of the element with the minimum value and the index of that element. If there are multiple elements, the indexes of all those elements should be printed.