Ch 5 Task 3 completed
Write a program with a static method that takes an array of integers and an integer as an argument. As a result, the method returns a reference to a new array, which is obtained from the original array (passed as the first argument to the method), if several initial elements are taken in it. The number of elements to be taken from the original array is determined by the second argument of the method. If the second argument of the method is greater than the length of the array, passed as the first argument, the method creates a copy of the original array and returns a reference to that copy.