Ch 5 Task 8 completed
Write a program with an overloaded static method. If two integers are passed as an argument to the method, then the result returns a reference to an integer array consisting of natural numbers, and the first and last number in this array are determined by the method arguments. If two character values ​​are passed as an arguments to the method, then the result returns a reference to an array consisting of a sequence of characters, and the first and last characters are determined by the method arguments.