Ch 5 Task 10 completed
Write a program with a static method that is passed text and an arbitrary number of symbolic arguments.  The result returns the text that is obtained by appending to the end of the source text (the first argument of the method) the character values ​​passed as arguments to the method.