Ch 6 Task 1 completed
Write a program with a class that has a private character field and three public methods. One method allows you to assign a value to a field. Another method, when called, returns a character code. The third method allows you to display the symbol (field value) and its code in the console window.