Ch 6 Task 2 completed
Write a program with a class that has two character fields and a method. It does not return a result and it has no arguments. When called, the method prints to the console window all the characters from the code table that are between the characters that are the values ​​of the fields of the object (from which the method is called).