Ch 6 Task 4 completed
Write a program with a class that contains a character and an integer field. The class must describe versions of the constructor with two arguments, as well as with one double argument. In the latter case, the real part of the argument specifies the character code (the value of the character field), and the fractional part (including tenths and hundredths) specifies the value of the integer field.