Ch 6 Task 5 completed
Write a program with a class that has a private integer field. A value is assigned to a field using a public method. An integer can be passed as an argument to the method, and the method can also be called without arguments. If the method is called without arguments, then the field gets a zero value. If the method is called with an integer argument, then that value is assigned to the field. However, if the value passed as an argument to the method is greater than 100, then the field's value is set to 100. Provide a constructor in your class that works in the same way as the method. Also, the class must have a method that allows you to check the value of the field.