Ch 6 Task 6
Write a program with a class that has two private integer fields (max and min). The value of the max field cannot be less than the value of the min field. Values ​​are assigned to fields using a public method. The method can be called with one or two integer arguments. When the method is called, the arguments are compared with the current values ​​of the fields. The largest value is assigned to the max field, and the smallest value is assigned to the min field. Provide a constructor that works in the same way as a method for assigning values ​​to fields. The class must also have a method that displays the values ​​of the object's fields in the console window.