Ch 6 Task 7 completed
Write a program with a class that has two fields: character and text. The class must have an overloaded method for assigning values ​​to fields. If the method is called with a character argument, the corresponding value is assigned to the character field. If the method is called with a text argument, then it determines the value of the text field. A character array can also be passed as an argument to the method. If the array has one element. then it defines the value of the character field. Otherwise (if there is more than one element), a text string is formed from the array character and assigned a value to the text field.