Ch 6 Task 9 completed
Write a program with a class that has static methods that can be passed an arbitrary number of integer arguments (or an integer array). Methods based on the passed arguments or an array allow you to calculate the largest value, the smallest value, and the average value from a set of numbers.