Ch 7 Task 2 completed
Write a program that has a static method that returns a text value as a result and receives text as an argument. The result of the method is the text passed as the argument, in which the words appear in reverse order. Words are blocks of text separated by spaces.