Ch 7 Task 3 completed
Write a program with a static method to compare text strings. Strings are compared character by character for a match. The rule for comparing characters is: two characters are considered the same if their codes differ by no more than one. Text strings match if they have the same characters (in the sense above).