Ch 7 Task 4 completed
Write a program with a static method that compares text strings. Text strings are compared in the following way: for each text value, a set of different letters is determined that are included in the text (the number of occurrences of a letter in the text does not matter). Text strings are considered equal if they contain the same set of letters.