Ch 7 Task 6 completed
Write a program with a static method with the text argument passed, and a result as a character array consisting of letters (excluding spaces and punctuation marks) that make up the entered text. If a letter occurs several times in the text, it must be represented by a single element in the array. The letters in the result array must be sorted alphabetically.