Ch 7 Task 7 completed
Write a program with a static method that emulates the Substring() method. The arguments to the static method are text and two integer arguments. As a result, the method returns a text string that consists of text characters (the first argument), starting from the position with the index (the second argument). The third argument to the static method specifies the number of characters to include in the substring.