Ch 7 Task 8 completed
Write a program with a class that has a text field. The value of the text field is assigned when the class object is created. Also, the class should have a method that allows you to insert a substring into the text from the text field. The substring to be inserted into the text is passed as arguments to the method, as well as the index of the position from which the insertion is performed. Redefine the ToString( ) method in the class so that it returns the value of the text from the text field.