Ch 7 Task 9 completed
Write a program with a class that has a text field and a character field. The value of the fields is assigned by the constructor. The class must have a method. returning an array of text strings. These strings are obtained by splitting the text field into substrings. The spliting indivator character is the character argument. Override the ToString( ) method in the class so that it returns the text with the value of the object's fields and the substrings the text splited to.