Ch 7 Task 10 completed
Write a program with a class that has a field - a reference to an integer array. When an object is created, the array is filled with random numbers. Override the ToString( ) method in the class so that the method returns a text string with the values ​​of the array elements and the average of the array elements.