Ch8Task3 conpleted
Write a program that has a class with two integer fields. Describe operator methods for this class that allow you to compare class objects for "less than" or "greater than". Assume that one object is less/greater than another if the sum of the squares of its field values is less than/greater than the sum of the squares of the field values of another object.