Ch8Task4 completed
Write a program that has a class with integer and text fields. Overload all comparison operators. The comparison for more or less is performed based on the comparison of the length of text values ​​(meaning the text fields of the objects being compared). When comparing for "greater than or equal to" or "less than or equal to", the values ​​of integer fields of objects are compared. When comparing for "equal" or "not equal", integer and text fields of objects are compared. Also suggest a way to override Equals( ) and GetHashCode( ) methods.