Ch8Task5 completed
Write a program that has a class with an integer field and a character field. Overload the "true" and "false" operators so that "true" is an object whose difference between the value of an integer field and a character code from a character field does not exceed 10. Use an object of this class (as a condition in a conditional operator) to to display a sequence of characters in a console window.