Ch8Task6 completed
Write a program that has a class with an integer field. Overload the &, |, true, and false operators so that the && and || operators can be used with class objects. The overload should be implemented so that an object is considered "true" if its numeric field value is 2, 3, 5, or 7. An object should be considered false if its numeric field value is less than 1 or greater than 10.