Ch8Task7 completed
Write a program that has a class with a text field. Describe operator methods in the class to perform type casting. You need to define the following transformation methods. When converting an object to an integer, the result returns the number of characters in the text field value. When converting an object to a character, the result is the first character in the text. When converting a number to an object, an object is created (and returned as a result) and its text field contains a text string of A characters. The number of characters in the text is determined by the converted number.