Ch8Task8 completed
Write a program that has a class with a field - a reference to an integer array. Describe operator methods for type casting in the class. You must implement the following type casting rules. When converting an object to text, a text string is returned with the values of the array elements. When converting an object to a number, the sum of the array elements is returned. When converting a number to an object, the result is a new object with an array size determined by the number being converted. The array in the object is filled with zeros.