Ch8Task10 completed
Write a program that has a class with a character field and an overload of the + and - operators. The operators must be overloaded so that applying the + operator to objects of the class results in text, which is the result of the concatenation of the character fields of the summed objects. When applying the - operator to objects of a class, the result must be an integer (the difference in character codes from the objects being subtracted). Suggest ways to overload cast operators so that the += and -= operators can be used with class objects.