Ch1Task1 completed
Write a program that contains an abstract base class with a protected field that is a reference to an integer array. The class must have a constructor with one argument (defines the size of the array and creates it), an integer property (the value is the size of the array), an abstract method (no arguments, does not return a result), and an indexer with an integer index (readable and writable). In the derived class, you need to describe the abstract method from the bass class so that it displays the contents of the array in the console. Define an indexer so that it can be used to read the value of an array element and assign a value to the array element.