Ch1Task2 completed
Write a program that contains a base class with a protected text field. The class must have a constructor with a text argument, a readable abstract integer property, a readable abstract indexer with an integer index, and an abstract method with no arguments that does not return a result. In a derived class: the property returns the number of characters in the text field as a result, the value of the expression with an indexed argument is the code of the character in the text, the method with the text argument assigns a new value to the field, and the method with no arguments displays the value of the field in the console window.