Ch1Task3 completed
Write a program that describes an interface with a method with no arguments that returns an integer as a result. Create two classes based on the interface. Each class must have an integer field. In the first class the method returns the sum of even numbers, in the second class the method returns the sum of odd numbers. The number of terms is determined by the field of the object from which the method is called. Test the method by accessing the class object through an object variable and through an interface variable.