Ch1Task7 completed
Write a program where a class is created based on two interfaces. The first interface has an integer property that is readable and writable. The second interface has a text property available for reading and writing. In each of the interfaces, the same (with the same names) methods are declared, without arguments, and not returning a result. In the class, describe the corresponding method that displays the property values in the console window.