Ch1Task8 completed
Write a program where a class is created based on two interfaces. Each interface has a read-only text property with the same name. In the class, perform a general (implicit) implementation of the property, as well as an explicit specialization of the property for each of the interfaces. Check the value of a property for a class object by referencing the object through the object variable and interface variables.