Ch1Task9 completed
Write a program that creates a class from an abstract base class and an interface. The abstract class has a field that is a reference to a protected character array. The class constructor is passed a text argument, on the basis of which a character array is created and filled. In an abstract class, describe a method that, given an integer argument, returns the value of the character at the corresponding index in the array. Also in the abstract class, an abstract method with two arguments (an integer and a character) must be declared that does not return a result. An interface must declare a method with the same name but with one text argument. Also, an indexer (with two accessors) with a character result and an integer index must be declared in the interface. Based on the abstract class and interface, you need to create a class. An abstract method from the base class must be overridden so that it assigns a value to an array element, a method from the interface must create a new array, and the indexer must provide access to the elements of the array.