Ch1Task10 completed
Write a program that contains an abstract class and two interfaces. The class must contain the declaration of an abstract property (with two accessors), an abstract indexer (with two accessors), and an abstract method. The same property, indexer, and method must be in interfaces. Based on the abstract class and interfaces, you need to create a class. In this class, you must make an explicit implementation for the property, indexer, and method for each of the interfaces. Test the operation of the property, indexer, and method by accessing the class object through an object variable and through interface variables.