Ch2Task1 completed
Write a program that declares a delegate for methods with two arguments (character and text) and an integer result. In the main class, it is necessary to describe two static methods. One static method returns the number of occurrences of a character (first argument) in a text string (second argument) as a result. Another method returns the index of the first occurrence of the character (first argument) in the text string (second argument) as a result, or -1 if the character does not occur in the text string. In the main method, create an instance of the delegate and use that instance to call each of the static methods.