Ch2Task2 completed
Write a program that declares a delegate for methods with a symbolic argument that do not return a result. Describe a class that should have a character field and a method that allows you to assign a value to a character field of an object. The method has one character argument and the method does not return a result. Create an array of objects of the given class. Create an instance of the delegate. In the call list of this delegate, you need to add references to the method (assigning a value to the character field) of each object from the array. Check the result of calling such a delegate instance.