Ch2Task5 completed
Write a program that uses delegates and anonymous methods (lambda expressions) to create an instance of a delegate that is called without arguments and returns a number from the Fibonacci sequence as a result - with each new call, we get a new number in the sequence. In the Fibonacci sequence, the first two numbers are equal to one, and each subsequent number is equal to the sum of the previous two.