Ch2Task6 completed
Write a program that uses delegates and anonymous methods (lambda expressions) to create an instance of a delegate that is called without arguments and returns a text value with the name of the day of the week ("Monday", "Tuesday" and so on until "Sunday"). Each time the instance is called again, the name of the next day of the week is returned as a result. After "Sunday" the result returns "Monday" and so on.