Ch2Task7 completed
Write a program containing a static method. The first argument to the static method is an integer array. The second argument to the static method is a reference to another method. An argument method must have an integer argument and must return an integer result. As a result, the static method returns an integer array. The elements of this array are calculated as the result of calling the argument method, if the values ​​of the elements from the argument array are passed to it. Suggest a mechanism for checking the functionality of this static method.