Ch2Task9 completed
Write a program containing a static method with three real (double) arguments. The result of a static method is a reference to a method that calculates a quadratic trinomial with coefficients determined by the arguments of the static method. In other words, if a static method is called with numeric arguments a, b, and c, then the result is a reference to a method that, given a double argument x, returns the value of the expression a*x*x+b*x+c as a result.