Ch6Task1 completed
Write a program that computes factorial, double factorial, and a Fibonacci number for a given integer value. A child thread is started to calculate each of the values. Factorial n! for the number n is calculated as the product of numbers from 1 to this number inclusive, that is, n! = 1 × 2 × 3 × ... × n. Double factorial n!! is calculated as the product of natural numbers "through one", that is, n!! = n × (n − 2) × (n − 4) × ... (the last factor is 2 or 1, depending on whether n is even or odd). In the Fibonacci sequence, the first two numbers are 1, and each next number is equal to the sum of the previous two (the numbers are 1, 1, 2, 3, 5, 8, 13, 21, and so on).