Ch6Task3 completed
Write a program in which an integer variable on the main thread receives a random value at regular intervals. Two child threads periodically (at regular intervals) check the value of the variable. The first thread checks if the value of the variable is odd, and the second thread checks if the value of the variable is divisible by 3. If the check is successful, the corresponding thread prints a message to the console window.