Ch6Task4 completed
Write a program in which two threads fill an integer array. One thread fills the array from beginning to end with numbers that are powers of two (2^0, 2^1, 2^2, and so on). Another thread fills the array from end to beginning with numbers that are powers of a triple (3^0, 3^1, 3^2, and so on).