Ch6Task5 completed
Write a program containing a class with two fields: one is a reference to an integer array, and the second field is a reference to a character array. Create an object of the class, and also run two child threads. One child thread must fill the object's character array, and the second child thread must fill the object's integer array. Suggest a filling method yourself (for example, random characters and numbers).