Ch6Task7 completed
Write a program that creates a two-dimensional integer array. The array is filled line by line: the first line is filled with numbers that are powers of two (2^0, 2^1, 2^2, and so on), the second line is filled with powers of three (3^0, 3^1, 3^2, and so on) , etc. A child thread is started to fill each row. Object variables for threads should be organized as an array.